+++ CHANGELOG.md
@@ -17,7 +17,7 @@ adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 ## [2021.2.3.28] - 2021-06-28
 
 ### Added
-- when creating downtime, adding userId and orderId, if valid
+- when creating downtime, adding userId and orderRecordId, if valid
 
 ## [2021.2.3.14] - 2021-06-14